Add a workaround for metadata encryption.
Add a workaround for is_ice_enabled to return 0 if the metadata
partition is present to support metadata encryption.

This is because HW FDE is a compile time flag and would go on to
encrypt the whole disk even for metadata encryption if this does
not return 0. Hence, checking for metadata partition.

Change-Id: I0b08aaf4fbe5b15acb3e8963e16e75b9ad515d09
Signed-off-by: Shivaprasad Hongal <shongal@codeaurora.org>